refactoring into multi schema_module + basic db write accomplished
