+++ basic_engine/tools/perf_test.py
@@ -6,6 +6,9 @@ sys.path.append('../')
 from src.main import Game
 from utils import convert_pgn_to_a1
 
+from schema_module import TimeAnalysisSchema
+from schema_module import TurnAttributeSchema
+
 
 
 # Different Experiments --------------------------------------------------
@@ -561,260 +564,6 @@ def analysis3(s_tests, s_instructions, **kwargs):
 import json
 import pprint
 
-class TurnAttributeSchema:
-    
-    def __init__(self):
-        self.s_instructions = None
-        self.s_pgn_instructions = None
-        
-        self.num_available = None
-        self.num_player_pieces = None
-
-    def load_instructions(self, insturctions, b_pgn=True):
-        ''' load the insturctions and note whether they are boolean '''
-        if b_pgn:
-            self.s_pgn_instructions = insturctions
-        else:
-            self.s_instructions = insturctions
-
-    def data_from_gamelog(self, gameLog):
-        ''' input a gameLog object to extract trun attributes '''
-        self.num_available = gameLog.get_log_num_available()
-        self.num_player_pieces = gameLog.get_log_num_player_pieces()
-        
-    
-    def create_data(self, **kwargs):
-        ''' run a game using instructions and pass the gameLog into'''
-
-        if self.s_instructions is not None:
-            game = Game(s_instructions=self.s_instructions
-                        ,b_log_full=True)            
-        elif self.s_pgn_instructions is not None:
-            game = Game(s_pgn_instructions=self.s_pgn_instructions
-                        ,b_log_full=True)
-        else:
-            print 'could not find instructions to run create_data()'
-            return None
-
-        game.play()
-
-        self.data_from_gamelog(game.get_gamelog())
-
-
-    def get_data(self,**kwargs):
-        ''' return data asa a dict '''
-        data = {}
-        
-        if self.s_instructions is not None:
-            data['s_instructions'] = self.s_instructions
-        else:
-            data['s_instructions'] = self.s_pgn_instructions
-
-        data['num_available'] = self.num_available
-        data['num_player_pieces'] = self.num_player_pieces
-        
-        return data
-
-
-class TimeAnalysisSchema:
-    '''
-    Games - a list of instructions: [ "1. c4...", "1. e4...", ...]
-    for each game: {
-                  
-                  meta_analysis: {
-                      // these hold must be held constant for different trials to make sense
-                      'algo_style':     : 'baseline_yk'
-                      'analysis_type'   : 'analysis2'   //by turn times 
-                        }
-                  
-                  log: {
-                       // these are "X" in regression
-                      'num_avaialble' : [20,22,26,...,0]
-                      'king_moves'    : [0,0,2,...,0]    
-                          ...
-                      }
-                  
-                  trials: [
-                  
-                    [trial]: {  //unnamed as the dict is simply an element in a list
-                        
-                        trial_meta: {
-                            // these are used to aggregate separate runs
-                            'N'             : 250        //number of trials run and aggregated
-                            'data-time-run' : 2018-6-13-13-12-22-345  //runs strated here
-                            }
-
-                        trial_data: {
-                            // this is "Y" in regression
-                            avg-turn-time: [0.028, 0.031, 0.051, ..., 0.288]  //index by turn-num
-                            }
-                          }
-                    , trial: {...}
-                    , trial: {...}
-                    ,...
-                    ]
-                  }
-    '''
-    
-    def __init__(self):
-        
-        self.log = None
-        
-        self.meta_analysis = None
-        self.algo_style = None
-        self.analysis_type = None
-
-        self.trials = []
-
-        self.trial = None
-        
-        self.trial_meta = None
-        self.N = None
-        self.date_time_run = None
-
-        self.trial_data = None
-
-    
-    def get_all(self):
-        
-        temp = {}
-        temp['log'] = self.log
-        temp['meta_analysis'] = self.meta_analysis
-        temp['trials'] = copy.deepcopy(self.trials)
-        
-        return copy.deepcopy(temp)
-
-    def set_meta_analysis(self, algo_style, analysis_type):
-        
-        self.algo_style = algo_style
-        self.analysis_type = analysis_type
-
-        self.meta_analysis = {}
-        self.meta_analysis['algo_style'] = self.algo_style
-        self.meta_analysis['analysis_type'] = self.analysis_type
-        
-
-    def set_trial_meta(self, N, **kwargs):
-        
-        self.N = N
-        self.date_time_run = time()
-        
-        self.trial_meta = {}
-        self.trial_meta['N'] = self.N
-        self.trial_meta['data_time_run'] = self.date_time_run
-
-    def set_log(self, log_data):
-        self.log = copy.deepcopy(log_data)
-
-    @staticmethod
-    def aggregate_y(data):
-        ''' average each i-th element together'''
-        try:
-            assert len(data) > 1
-            len0  = len(data[0])
-            assert all(map(lambda data_i: len(data_i) == len0, data))
-        except AssertionError:
-            print 'data fed to aggregate_y is of different length; cant aggregate that.'
-            return None
-        try:
-            temp = []
-            n = len(data)
-            j = len(data[0])
-            for _j in range(j):
-                sum_elems = sum(map(lambda d: d[_j], data))
-                avg_elems = float(sum_elems) / float(n)
-                temp.append(avg_elems)
-            return temp
-        except:
-            print 'failed to aggregate, but did not catch in exception'
-            return None
-
-    @staticmethod
-    def weighted_avg(list_n, list_list_y):
-        
-        num_elems = len(list_n)
-        num_turns = len(list_list_y[0])
-        total_n = sum(list_n)
-        
-        cum_sum_y = [0 for j in range(num_turns)]
-        
-        for i in range(num_elems):
-            weighted_y = map(lambda elem: elem * float(list_n[i]), list_list_y[i])
-            cum_sum_y = [cum_sum_y[j] + weighted_y[j] for j in range(num_turns)]
-
-        avg_y = [float(cum_sum_y[j]) / float(total_n) for j in range(num_turns)]
-
-        return avg_y
-
-        
-    def weighted_avg_trials(self):
-        
-        list_n = []
-        list_list_y = []
-        
-        for trial in range(len(self.trials)):
-            
-            list_n.append(trial['trial_meta']['N'])
-            list_list_y.append(trial['trial_data'])
-
-        return self.weighted_avg(list_n, list_list_y)
-
-
-    def set_trial_data(self, analysis_results):
-        ''' set results of analysis to temp var: trial_data'''
-        self.trial_data = self.aggregate_y(analysis_results['turn_time'])
-
-
-    def add_trial(self):
-        ''' append trial_meta and trial_data onto trials '''
-        self.trial = {}
-        self.trial['trial_meta'] = self.trial_meta
-        self.trial['trial_data'] = self.trial_data
-
-        self.trials.append(self.trial)
-        
-
-
-    def to_json(self, data_dir=None, data_fn=None):
-        ''' return json string or write to json file '''
-        
-        tas = self.get_all()
-        
-        if data_dir is None:
-            return json.dumps(tas)
-    
-        #write to output
-        data_dir = '../data/perf/'
-        data_fn = 'demo.tas'
-        try:
-            with open(data_dir + data_fn, 'w') as f:
-                json.dump(tas, f)
-        except:
-            print 'failed to write file output'
-            return None
-        print 'wrote output to: ', data_dir, data_fn
-
-    
-    def from_json(self, path_fn="../data/perf/demo.tas"):
-        ''' load the TAS from json file'''
-        try:
-            with open(path_fn, "r") as f:
-                d_tas = json.load(f)
-        except:
-            print 'could not load json from: ', str(path_fn)
-            return None
-
-        try:
-            self.log = d_tas['log']
-            self.meta_analysis = d_tas['meta_analysis']
-            self.trials = d_tas['trials']
-
-        except:
-            print 'could not convert json to TAS'
-            return None
-
-        return 0
-
 
 def one_analysis(   s_instructions
                     ,b_pgn=True
@@ -823,6 +572,7 @@ def one_analysis(   s_instructions
                     ,b_noisy=False
                     ,b_return_tas=False
                     ,b_write_out=False
+                    ,b_build_x=True
                     ):
     
     ''' run analysis2 to create Y-data, run create_data() to create X-data()'''
@@ -834,13 +584,16 @@ def one_analysis(   s_instructions
                                         ,analysis_type = 'analysis2'
                                         )
     
-    #Build X
-    turn_attributes = TurnAttributeSchema()
-    turn_attributes.load_instructions(s_instructions, b_pgn=b_pgn)
-    turn_attributes.create_data()
+    
+    if b_build_x:
+        
+        #Build X
+        turn_attributes = TurnAttributeSchema()
+        turn_attributes.load_instructions(s_instructions, b_pgn=b_pgn)
+        turn_attributes.create_data()
 
-    #Set X
-    analysis_schema.set_log(turn_attributes.get_data())
+        #Set X
+        analysis_schema.set_log(turn_attributes.get_data())
 
     #Set Y-(trial)-meta
     analysis_schema.set_trial_meta(N = n)
@@ -873,6 +626,8 @@ def one_analysis(   s_instructions
         return analysis_schema
 
 
+from db_module import DBDriver
+
 
 def batch_analyze(   input_fn="GarryKasparovGames.txt"
                     ,output_fn="demo_batch.tas"
@@ -886,12 +641,24 @@ def batch_analyze(   input_fn="GarryKasparovGames.txt"
     ''' analyze mulitple games and log them '''
 
     
-    # FILESYSTEM Params
+    #PARAMS
+    b_input_filesystem = True
+    b_output_filesystem = False
+    b_input_db = False
+    b_output_db = True
+
     INPUT_DATA_DIR = "../data/"
     OUTPUT_DATA_DIR = "../data/perf/"
     
-    input_data = INPUT_DATA_DIR + input_fn
-    output_data = OUTPUT_DATA_DIR + output_fn
+    if (output_fn is None) or (input_fn is None):
+        db = DBDriver(b_noisy=False)
+
+    b_insert = True if input_fn is not None else False  #insert or update
+
+    if input_fn is not None:
+        input_data = INPUT_DATA_DIR + input_fn
+    if output_fn is not None:
+        output_data = OUTPUT_DATA_DIR + output_fn
     
     #Read in Data
     with open(input_data, "r") as f:
@@ -912,10 +679,16 @@ def batch_analyze(   input_fn="GarryKasparovGames.txt"
                             ,b_noisy=False
                             ,b_return_tas=True
                             ,b_write_out=False
+                            ,b_build_x=b_insert
                             )
 
         key_name = input_fn + "-" + str(i+1)
-        results[key_name] = ret.get_all()
+        
+        if output_fn is not None:
+            results[key_name] = ret.get_all()
+        else:
+            results[key_name] = ret.to_json()
+            
         if b_noisy:
             print '\n'+ key_name + '\n'
     
@@ -924,8 +697,17 @@ def batch_analyze(   input_fn="GarryKasparovGames.txt"
         print 'done with batch_analyze'
 
     if b_write_out:
-        with open(output_data, "w") as f:
-            json.dump(results, f)
+        
+        if output_fn is not None:
+            with open(output_data, "w") as f:
+                json.dump(results, f)
+
+        if output_fn is None:
+            for _k in results.keys():
+                db.add_basic_record(s_tas = results[_k] ,tas_id = _k)
+
+            db.close_conn()
+            
 
 
         
@@ -941,6 +723,7 @@ def batch_analyze(   input_fn="GarryKasparovGames.txt"
 # > python perf_test.py --gameinitdemo
 # > python perf_test.py --batchdemo
 # > python perf_test.py --batchdemosave
+# > python perf_test.py --batchdemodb
 # > python perf_test.py --singledemo
 
 if __name__ == "__main__":
@@ -957,6 +740,7 @@ if __name__ == "__main__":
     ap.add_argument("--pgndemo", action="store_true")
     ap.add_argument("--batchdemo", action="store_true")
     ap.add_argument("--batchdemosave", action="store_true")
+    ap.add_argument("--batchdemodb", action="store_true")
     ap.add_argument("--singledemo", action="store_true")
 
     args = vars(ap.parse_args())
@@ -1068,10 +852,21 @@ if __name__ == "__main__":
     if args["batchdemosave"]:
         batch_analyze(max_lines=2, n=2, b_noisy=False, b_write_out=True)
 
+    if args["batchdemodb"]:
+        db = DBDriver()
+        db.drop_table_basic_tas()
+        db.close_conn()
+        batch_analyze(max_lines=2, n=2, b_noisy=False, b_write_out=True
+                        ,output_fn=None)
+        db = DBDriver()
+        ret = db.select_all_basic()
+        print ret
+
     if args["singledemo"]:
         s_instructions = "1. d4 e6 2. Nf3 Nf6"
         # s_instructions = "1. d4 e6 2. Nf3 Nf6 3. c4 Bb4+ 4. Nc3 b6 5. Qb3 Qe7 6. Bf4 d5 7. e3 Bb7 8. a3 Bxc3+ 9. Qxc3 O-O 10. Be2 dxc4 11. Qxc4 Rc8 12. O-O Ba6 13. Qc2 Bxe2 14. Qxe2 c5 15. Rac1 Nbd7 16. Qa6 h6 17. h3 Qe8 18. Bh2 cxd4 19. Nxd4 Nc5 20. Qe2 Qa4 21. Rc4 Qe8 22. Rfc1 a5 23. f3 a4 24. e4 Nfd7 25. Nb5 Qe7 26. Qe3 Qf6 27. e5 Qg6 28. Nd6 Rd8 29. Rg4 Qh7 30. Bf4 Kf8 31. Rd1 f5 32. exf6 Nxf6 33. Rh4 Nd5 34. Qe5 Nxf4 35. Rxf4+ Kg8 36. Rfd4 Rd7 37. Ne4 Rxd4 38. Rxd4 Qf5 39. Qxf5 exf5 40. Nxc5 bxc5 41. Rd5 Ra5 42. Rxf5 Rb5 43. Rf4 Rxb2 44. Rxa4 Ra2 45. h4 c4 46. Rxc4 Rxa3 47. Rc5 Ra4 48. h5 Ra2 49. Kh2 Rb2 50. Kh3 Rd2 51. g4 Rd1 52. Kg3 Rd4 53. Rc7 Ra4 54. Re7 Kf8 55. Re4 Ra5 56. Kf4 Kf7 57. Re5 Ra3 58. Ke4 Rb3 59. f4 Rb4+ 60. Kf5 Rb7 61. Rc5 Ra7 62. g5 hxg5 63. fxg5 g6+ 64. hxg6+ Kg7 65. Rc6 Ra5+ 66. Kg4 Ra1"
         one_analysis(s_instructions, b_noisy=True)
+        
 
 
     print 'script done.'
